% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bachelier_impvol.R
\name{BachelierImpvol}
\alias{BachelierImpvol}
\title{Calculate Bachelier model implied volatility}
\usage{
BachelierImpvol(
  type = "call",
  price,
  spot,
  forward = spot * exp((intr - divr) * texp),
  strike = forward,
  texp = 1,
  intr = 0,
  divr = 0
)
}
\arguments{
\item{type}{option type either 'call' or 'put'}

\item{price}{Price}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{strike}{strike price}

\item{texp}{time to expiry}

\item{intr}{interest rate}

\item{divr}{dividend rate}
}
\value{
implied vol
}
\description{
Calculate Bachelier model implied volatility
}
\examples{
spot <- 100
strike <- 100
texp <- 1.2
sigma <- 0.2
intr <- 0.05
price <- 20
vol <- FER::BachelierImpvol(price=price, spot=spot, strike=strike, texp=texp, intr=intr)
}
