% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes_impvol.R
\name{BlackScholesImpvol}
\alias{BlackScholesImpvol}
\title{Calculate Black-Scholes implied volatility}
\usage{
BlackScholesImpvol(
  price,
  strike = forward,
  spot,
  texp = 1,
  intr = 0,
  divr = 0,
  cpsign = 1,
  forward = spot * exp(-divr * texp)/df,
  df = exp(-intr * texp)
)
}
\arguments{
\item{price}{(vector of) option price}

\item{strike}{(vector of) strike price}

\item{spot}{(vector of) spot price}

\item{texp}{(vector of) time to expiry}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{cpsign}{call/put sign. 1 for call, -1 for put.}

\item{forward}{forward price. If given, \code{forward} overrides \code{spot}}

\item{df}{discount factor. If given, \code{df} overrides \code{intr}}
}
\value{
Black-Scholes implied volatility
}
\description{
Calculate Black-Scholes implied volatility
}
\examples{
spot <- 100
strike <- 100
texp <- 1.2
sigma <- 0.2
intr <- 0.05
price <- 20
vol <- FER::BlackScholesImpvol(price, strike, spot, texp, intr=intr)

}
\seealso{
\code{\link{BlackScholesPrice}}
}
