% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm_price.R
\name{CalcBsmPrice}
\alias{CalcBsmPrice}
\title{Calculate Black-Scholes-Merton option price}
\usage{
CalcBsmPrice(type = "call", spot, forward = spot * exp((r - div) * t.exp),
  strike = forward, t.exp = 1, r = 0, div = 0, sigma)
}
\arguments{
\item{type}{option type either 'call' or 'put'}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{strike}{strike price}

\item{t.exp}{time to expiry}

\item{r}{interest rate}

\item{div}{dividend rate}

\item{sigma}{volatility}
}
\value{
option price
}
\description{
Calculate Black-Scholes-Merton option price
}
\examples{
spot <- 100
strike <- seq(80,125,5)
t.exp <- 1.2
sigma <- 0.2
r <- 0.05
price <- phbsasp::CalcBsmPrice(spot=spot, t.exp = t.exp, sigma=sigma, strike=strike, r=r)
}

