% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{CEV}
\alias{CEV}
\title{Constant Elasticity Of Variance (CEV) model option price}
\usage{
CEV(
  strike,
  spot,
  forward = spot * exp((r - div) * t.exp),
  t.exp = 1,
  sigma = 0.01,
  beta = 0.5,
  r = 0,
  div = 0,
  type = "call"
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Constant Elasticity Of Variance (CEV) model option price
}
