% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev_price.R
\name{CevPrice}
\alias{CevPrice}
\title{Constant Elasticity Of Variance (CEV) model option price}
\usage{
CevPrice(
  strike = forward,
  spot,
  texp = 1,
  sigma,
  beta = 0.5,
  intr = 0,
  divr = 0,
  cpsign = 1,
  forward = spot * exp(-divr * texp)/df,
  df = exp(-intr * texp)
)
}
\arguments{
\item{strike}{(vector of) strike price}

\item{spot}{(vector of) spot price}

\item{texp}{(vector of) time to expiry}

\item{sigma}{(vector of) volatility}

\item{beta}{beta}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{cpsign}{call/put sign. 1 for call, -1 for put.}

\item{forward}{forward price. If given, forward overrides spot}

\item{df}{discount factor. If given, df overrides intr}
}
\value{
option price
}
\description{
Constant Elasticity Of Variance (CEV) model option price
}
\examples{
spot <- 100
strike <- seq(80,125,5)
texp <- 1.2
beta <- 0.5
sigma <- 2
price <- FER::CevPrice(strike, spot, texp, sigma, beta)

}
