% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bachelier_impvol.R
\name{BachelierImpvol}
\alias{BachelierImpvol}
\title{Calculate Bachelier model implied volatility}
\usage{
BachelierImpvol(
  type = "call",
  price,
  spot,
  forward = spot * exp((r - div) * t.exp),
  strike = forward,
  t.exp = 1,
  r = 0,
  div = 0
)
}
\arguments{
\item{type}{option type either 'call' or 'put'}

\item{price}{Price}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{strike}{strike price}

\item{t.exp}{time to expiry}

\item{r}{interest rate}

\item{div}{dividend rate}
}
\value{
implied vol
}
\description{
Calculate Bachelier model implied volatility
}
\examples{
spot <- 100
strike <- 100
t.exp <- 1.2
sigma <- 0.2
r <- 0.05
price <- 20
vol <- FER::BachelierImpvol(price=price, spot=spot, strike=strike, t.exp=t.exp, r=r)
}
