% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bachelier_price.R
\name{BachelierPrice}
\alias{BachelierPrice}
\title{Calculate Bachelier model option price}
\usage{
BachelierPrice(
  type = "call",
  spot,
  forward = spot * exp((intr - divr) * texp),
  strike = forward,
  texp = 1,
  intr = 0,
  divr = 0,
  sigma
)
}
\arguments{
\item{type}{option type either "call" or "put"}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{texp}{time to expiry}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{sigma}{volatility}

\item{price}{Price}
}
\value{
option price
}
\description{
Calculate Bachelier model option price
}
\examples{
spot <- 100
strike <- seq(80,125,5)
texp <- 1.2
sigma <- 0.2
intr <- 0.05
price <- FER::BachelierPrice(spot=spot, texp = texp, sigma=sigma, strike=strike, intr=intr)
}
