% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes_price.R
\name{BlackScholesPrice}
\alias{BlackScholesPrice}
\title{Calculate Black-Scholes option price}
\usage{
BlackScholesPrice(
  type = "call",
  spot,
  forward = spot * exp((intr - divr) * texp),
  strike = forward,
  texp = 1,
  intr = 0,
  divr = 0,
  sigma
)
}
\arguments{
\item{type}{option type either 'call' or 'put'}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{strike}{strike price}

\item{texp}{time to expiry}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{sigma}{volatility}
}
\value{
option price
}
\description{
Calculate Black-Scholes option price
}
\examples{
spot <- 100
strike <- seq(80,125,5)
texp <- 1.2
sigma <- 0.2
intr <- 0.05
price <- FER::BlackScholesPrice(spot=spot, texp = texp, sigma=sigma, strike=strike, intr=intr)
}
