% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cev.R
\name{CEV}
\alias{CEV}
\title{Constant Elasticity Of Variance (CEV) model option price}
\usage{
CEV(
  strike,
  spot,
  forward = spot * exp((intr - divr) * texp),
  texp = 1,
  sigma = 0.01,
  beta = 0.5,
  intr = 0,
  divr = 0,
  type = "call"
)
}
\arguments{
\item{strike}{strike price}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{texp}{time to expiry}

\item{sigma}{volatility}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{type}{option type either 'call' or 'put'}
}
\value{
option price
}
\description{
Constant Elasticity Of Variance (CEV) model option price
}
