% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_hagan2002.R
\name{SabrHagan2002}
\alias{SabrHagan2002}
\title{Hagan approximation for the SABR model}
\usage{
SabrHagan2002(
  strike,
  spot,
  forward = spot * exp((intr - divr) * texp),
  texp = 1,
  sigma = 0.01,
  vov = 0,
  rho = 0,
  beta = 1,
  intr = 0,
  divr = 0,
  type = "BlackScholes"
)
}
\arguments{
\item{strike}{strike price}

\item{spot}{current stock price}

\item{forward}{forward stock price}

\item{texp}{time to expiry}

\item{sigma}{volatility}

\item{intr}{interest rate}

\item{divr}{dividend rate}

\item{type}{'BlackScholes', 'call', 'put'}
}
\value{
Black-Scholes volatility (type="BlackScholes") or option price ("call" or "put")
}
\description{
Hagan approximation for the SABR model
}
