% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabr_hagan2002.R
\name{SabrHagan2002}
\alias{SabrHagan2002}
\title{Hagan approximation for the SABR model}
\usage{
SabrHagan2002(
  strike,
  spot,
  forward = spot * exp((r - div) * t.exp),
  t.exp = 1,
  sigma = 0.01,
  vov = 0,
  rho = 0,
  beta = 1,
  r = 0,
  div = 0,
  type = "BlackScholes"
)
}
\arguments{
\item{type}{'BlackScholes', 'call', 'put'}
}
\value{
Black-Scholes volatility (type="BlackScholes") or option price ("call" or "put")
}
\description{
Hagan approximation for the SABR model
}
